requirements.txt


# ----------------------------------------------------------
# Pre-commit
# ----------------------------------------------------------
pre-commit>=3.6,<4.0

# ----------------------------------------------------------
# Linting / Formatting
# ----------------------------------------------------------
ruff>=0.4,<0.6          # lint + format (substitui flake8, pylint em muitos casos)
black>=24.3,<25.0       # formatação consistente
isort>=5.13,<6.0        # ordenação de imports (caso não use ruff-format)

# ----------------------------------------------------------
# Testes (pytest)
# ----------------------------------------------------------
pytest>=8.0,<9.0

# ----------------------------------------------------------
# Testes estruturais de Airflow
# ----------------------------------------------------------
apache-airflow>=2.8,<3.0
pendulum>=3.0,<4.0          # datas (usado no Airflow)

# ----------------------------------------------------------
# Segurança
# ----------------------------------------------------------
detect-secrets>=1.5,<2.0

# ----------------------------------------------------------
# BDD / Gherkin (opcional – só se usar)
# ----------------------------------------------------------
behave>=1.2,<2.0
pytest-bdd>=7.1,<8.0

# ----------------------------------------------------------
# Qualidade adicional (opcional)
# ----------------------------------------------------------
mypy>=1.9,<2.0              # tipagem estática (se adotado)
types-requests              # stubs de tipos (exemplo)

# ----------------------------------------------------------
# Utilidades
# ----------------------------------------------------------
pyyaml>=6.0,<7.0            # leitura de YAML (configs, dynamic DAGs)
python-dotenv>=1.0,<2.0     # variáveis de ambiente em dev
